refactor(overlapping-trips): move block overlap check to ServiceValidator
